refactor: 투두랑 캘린더 삭제 로직을 스케줄로 통합
