+++ src/main/java/com/clush/assignment/domain/schedule/repository/ScheduleRepository.java
@@ -1,0 +1,7 @@
+package com.clush.assignment.domain.schedule.repository;
+
+import com.clush.assignment.domain.schedule.entity.Schedule;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
+}
